Migrate enum to TS
